uint32_t -> unsigned long in object.h

+++ src/object.h
@@ -4,29 +4,29 @@
 struct object {
 	union {
 		float x;
-		uint32_t xi;
+		unsigned long xi;
 	};
 	union {
 		float y;
-		uint32_t yi;
+		unsigned long yi;
 	};
 	union {
 		float z;
-		uint32_t zi;
+		unsigned long zi;
 	};
 	float speed;
 	int angle;
 	union {
 		float size_x;
-		uint32_t size_xi;
+		unsigned long size_xi;
 	};
 	union {
 		float size_y;
-		uint32_t size_yi;
+		unsigned long size_yi;
 	};
 	union {
 		float size_z;
-		uint32_t size_zi;
+		unsigned long size_zi;
 	};
 	float circle; /* used for rough estimations */
 	struct square *square;
